+++ streampark-flink/streampark-flink-kubernetes/src/main/scala/org/apache/streampark/flink/kubernetes/ingress/IngressStrategyV1beta1.scala
@@ -17,33 +17,38 @@
 
 package org.apache.streampark.flink.kubernetes.ingress
 
-import io.fabric8.kubernetes.api.model.OwnerReferenceBuilder
+import io.fabric8.kubernetes.api.model.{IntOrString, OwnerReferenceBuilder}
+import io.fabric8.kubernetes.api.model.networking.v1beta1.IngressBuilder
 import io.fabric8.kubernetes.client.DefaultKubernetesClient
+import org.apache.flink.client.program.ClusterClient
 
 import scala.collection.JavaConverters._
 import scala.language.postfixOps
 import scala.util.{Success, Try}
 
-class IngressStrategyV1beta1 extends BaseIngressStrategy {
+class IngressStrategyV1beta1 extends IngressStrategy {
+
+  override def ingressUrlAddress(
+      nameSpace: String,
+      clusterId: String,
+      clusterClient: ClusterClient[_]): String = {
+    val client = new DefaultKubernetesClient
+    // for kubernetes 1.19-
+    val fromV1beta1 =
+      Option(client.network.v1beta1.ingresses.inNamespace(nameSpace).withName(clusterId).get)
+        .map(ingress => ingress.getSpec.getRules.get(0))
+        .map(rule => rule.getHost -> rule.getHttp.getPaths.get(0).getPath)
+    Try(
+      fromV1beta1
+        .map { case (host, path) => s"https://$host$path" }
+        .getOrElse(clusterClient.getWebInterfaceURL)
+    ).getOrElse(throw new RuntimeException("[StreamPark] get ingressUrlAddress error."))
+  }
+
   override def configureIngress(domainName: String, clusterId: String, nameSpace: String): Unit = {
     Try(new DefaultKubernetesClient) match {
       case Success(client) =>
-        val annotMap = Map[String, String](
-          "nginx.ingress.kubernetes.io/rewrite-target" -> "/$2",
-          "nginx.ingress.kubernetes.io/proxy-body-size" -> "1024m",
-          "nginx.ingress.kubernetes.io/configuration-snippet" -> ("rewrite ^(/" + clusterId + ")$ $1/ permanent;")
-        )
-        val labelsMap = Map[String, String](
-          "app" -> clusterId,
-          "type" -> "flink-native-kubernetes",
-          "component" -> "ingress")
-
-        val deployment = client
-          .apps()
-          .deployments()
-          .inNamespace(nameSpace)
-          .withName(clusterId)
-          .get()
+        val deployment = getDeployment(nameSpace, clusterId, client)
 
         val deploymentUid = if (deployment != null) {
           deployment.getMetadata.getUid
@@ -62,11 +67,11 @@ class IngressStrategyV1beta1 extends BaseIngressStrategy {
           .withBlockOwnerDeletion(true)
           .build()
 
-        val ingress = new io.fabric8.kubernetes.api.model.networking.v1.IngressBuilder()
+        val ingress = new IngressBuilder()
           .withNewMetadata()
           .withName(clusterId)
-          .addToAnnotations(annotMap.asJava)
-          .addToLabels(labelsMap.asJava)
+          .addToAnnotations(buildIngressAnnotations(clusterId).asJava)
+          .addToLabels(buildIngressLabels(clusterId).asJava)
           .addToOwnerReferences(ownerReference)
           .endMetadata()
           .withNewSpec()
@@ -75,34 +80,23 @@ class IngressStrategyV1beta1 extends BaseIngressStrategy {
           .withNewHttp()
           .addNewPath()
           .withPath(s"/$nameSpace/$clusterId/")
-          .withPathType("ImplementationSpecific")
           .withNewBackend()
-          .withNewService()
-          .withName(s"$clusterId-rest")
-          .withNewPort()
-          .withName("rest")
-          .endPort()
-          .endService()
+          .withServiceName(s"$clusterId-rest")
+          .withServicePort(new IntOrString("rest"))
           .endBackend()
           .endPath()
           .addNewPath()
           .withPath(s"/$nameSpace/$clusterId" + "(/|$)(.*)")
-          .withPathType("ImplementationSpecific")
           .withNewBackend()
-          .withNewService()
-          .withName(s"$clusterId-rest")
-          .withNewPort()
-          .withName("rest")
-          .endPort()
-          .endService()
+          .withServiceName(s"$clusterId-rest")
+          .withServicePort(new IntOrString("rest"))
           .endBackend()
           .endPath()
           .endHttp()
           .endRule()
           .endSpec()
           .build()
-        client.network.v1.ingresses().inNamespace(nameSpace).create(ingress)
-
+        client.network.ingress.inNamespace(nameSpace).create(ingress)
       case _ =>
     }
   }
